botGuara.py/requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==37.0.4
discord.py==1.7.3
h11==0.13.0
idna==3.3
multidict==6.0.2
mysql-connector-python==8.0.29
outcome==1.2.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.27.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
socketIO-client==0.7.2
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.8
webdriver-manager==3.8.2
websocket-client==1.3.1
wsproto==1.1.0
yarl==1.7.2
youtube-dl==2021.12.17
youtube-search==2.1.0
